+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+#  
+
+## 0.0.3
+
+* TODO: Add to get related Calendar Reminder through ID
+
+## 0.0.2
+
+* TODO: Improve relevant IOS Android code
+
 ## 0.0.1
 
 * TODO: Describe initial release.